refactor(sicklegraph): consolidate files and update project structure
This commit consolidates various files and updates the project structure to improve maintainability. Key changes include:
- Moved `sicklegraph-plugin` files to the root `src` directory
- Deleted redundant test files and configurations
- Updated package.json to reflect plugin-specific dependencies
- Added new environment schema and example file
- Updated README with deployment guide and image requirements
