--- .env.example
@@ -1,46 +1,0 @@
-# OpenAI Configuration (required if using the Openai LLM Plugin)
-OPENAI_API_KEY=
-
-# Anthropic Configuration (required if using the Anthropic LLM Plugin)
-ANTHROPIC_API_KEY=
-
-# Fill these out if you want to use Discord (required if using the Discord plugin)
-DISCORD_APPLICATION_ID=
-DISCORD_API_TOKEN=
-
-# Fill these out if you want to use Postgres (required if using custom postgres url like neon db or railways)
-POSTGRES_URL=
-
-# Fill these out if you want to use Telegram (required if using the Telegram plugin)
-TELEGRAM_BOT_TOKEN=
-
-# Fill these out if you want to use Twitter (required if using the Twitter plugin)
-TWITTER_USERNAME=
-TWITTER_PASSWORD=
-TWITTER_EMAIL=
-TWITTER_ENABLE_POST_GENERATION=
-
-# Fill these out if you want to use EVM (required if using any wallet plugin)
-EVM_PRIVATE_KEY=
-EVM_CHAINS=mainnet,sepolia,base,arbitrum,polygon
-EVM_PROVIDER_URL=
-
-# Fill these out if you want to use Solana
-SOLANA_PUBLIC_KEY=
-SOLANA_PRIVATE_KEY=
-BIRDEYE_API_KEY=
-
-# Ollama Configuration (required if using the OLLAMA plugin)
-OLLAMA_API_ENDPOINT=
-OLLAMA_MODEL=
-USE_OLLAMA_EMBEDDING=
-OLLAMA_EMBEDDING_MODEL=
-OLLAMA_SMALL_MODEL=
-OLLAMA_MEDIUM_MODEL=
-OLLAMA_LARGE_MODEL=
-
-# StudioLM Configuration (required if using the StudioLM plugin)
-STUDIOLM_SERVER_URL=
-STUDIOLM_SMALL_MODEL=
-STUDIOLM_MEDIUM_MODEL=
-STUDIOLM_EMBEDDING_MODEL=
