+++ env.example
@@ -1,0 +1,16 @@
+# OpenAI Configuration (required if using the Openai LLM Plugin)
+OPENAI_API_KEY=
+
+# Groq Configuration
+GROQ_API_KEY=                               # Starts with gsk_
+SMALL_GROQ_MODEL=                           # Default: llama-3.1-8b-instant
+MEDIUM_GROQ_MODEL=                          # Default: llama-3.3-70b-versatile
+LARGE_GROQ_MODEL=                           # Default: llama-3.2-90b-vision-preview
+EMBEDDING_GROQ_MODEL=                       # Default: llama-3.1-8b-instant
+
+# Kuzu DB
+KUZU_DB_PATH=
+
+# NCBI API
+NCBI_API_KEY=""
+NCBI_BASE_URL="https://eutils.ncbi.nlm.nih.gov/entrez/eutils/"
