+++ images/README.md
@@ -1,0 +1,24 @@
+# Required Images for ElizaOS Plugins
+
+Please add the following required images to this directory:
+
+## logo.jpg
+
+- **Size**: 400x400px square
+- **Max size**: 500KB
+- **Purpose**: Main logo for your plugin displayed in the registry and UI
+
+## banner.jpg
+
+- **Size**: 1280x640px (2:1 aspect ratio)
+- **Max size**: 1MB
+- **Purpose**: Banner image for your plugin displayed in the registry
+
+## Guidelines
+
+- Use clear, high-resolution images
+- Keep file sizes optimized
+- Follow the ElizaOS brand guidelines
+- Include alt text in your documentation for accessibility
+
+These files are required for registry submission. Your plugin submission will not be accepted without these images.
