+++ src/character/sicklegraph-agent.ts
@@ -7,13 +7,12 @@ export const character: Character = {
         ...(process.env.OPENAI_API_KEY ? ["@elizaos/plugin-openai"] : []),
         ...(process.env.ANTHROPIC_API_KEY ? ["@elizaos/plugin-anthropic"] : []),
         ...(!process.env.OPENAI_API_KEY && !process.env.ANTHROPIC_API_KEY
-          ? ["@elizaos/plugin-local-ai"]
-          : []),
-        
+            ? ["@elizaos/plugin-local-ai"] : []),
+
         ...(process.env.DISCORD_API_TOKEN ? ["@elizaos/plugin-discord"] : []),
         ...(process.env.TWITTER_USERNAME ? ["@elizaos/plugin-twitter"] : []),
         ...(process.env.TELEGRAM_BOT_TOKEN ? ["@elizaos/plugin-telegram"] : []),
-      ],
+    ],
     settings: {
         secrets: {},
     },
