--- src/sicklegraph-plugin/index.ts
@@ -1,28 +1,0 @@
-import { IAgentRuntime, logger, Plugin } from "@elizaos/core";
-import { SickleGraphRoutes } from "./routes";
-import { ResearchAssistantAction } from "./actions/research-assistant";
-import { SickleGraphService } from "./services/sickle-graph-service";
-
-/**
- * SickleGraph plugin for ElizaOS
- */
-export const SickleGraphPlugin: Plugin = {
-    name: 'SickleGraph',
-    description: 'Biomedical knowledge graph for gene therapy research',
-    
-    actions: [ResearchAssistantAction],
-    routes: SickleGraphRoutes,
-
-    services: [],
-    providers: [],
-    evaluators: [],
-
-    init: async (config: Record<string, string>, runtime: IAgentRuntime) => {
-        logger.info('Initializing SickleGraph plugin');
-        // Warm up the service
-        const service = await runtime.getService<SickleGraphService>(
-            SickleGraphService.serviceType
-        );
-        logger.info('SickleGraph plugin ready');
-    },
-}
